+++ src/from_scratch/Makefile
@@ -1,15 +1,15 @@
 books:
-	poetry run python feature_pipeline/data_extraction.py
+	poetry run python data_extraction.py
 
 process:
-	poetry run python feature_pipeline/preprocessing.py --use_spacy --describe
+	poetry run python processing.py --use_spacy --describe
 
 tokens:
-	poetry run python feature_pipeline/tokenizer.py
+	poetry run python tokenizer.py
 
 embeddings:
-	poetry run python feature_pipeline/embeddings.py 
+	poetry run python embeddings.py 
 
 
 search:
-	poetry run python feature_pipeline/semantic_search.py 
+	poetry run python semantic_search.py
