+++ src/from_scratch/chunking.py
@@ -6,10 +6,10 @@ import json
 
 from tqdm import tqdm
 
-from setup.config import config
-from setup.paths import CHUNK_DETAILS_DIR
-from feature_pipeline.data_extraction import Book
-from feature_pipeline.reading import remove_new_line_marker
+from config import config
+from data_extraction import Book
+from paths import CHUNK_DETAILS_DIR
+from reading import remove_new_line_marker
 
 
 def perform_sentence_chunking(book: Book, details_of_all_pages: dict[str, str|int]) -> list[dict[str, str|int]]:
