+++ src/from_scratch/reading.py
@@ -15,10 +15,10 @@ from pathlib import Path
 from loguru import logger
 from pymupdf import Document
 
-from setup.config import find_non_core_pages 
-from feature_pipeline.data_extraction import Book
-from setup.paths import CLEANED_TEXT, make_data_directories
-from feature_pipeline.segmentation import get_tokens_with_spacy, segment_with_spacy, add_spacy_pipeline_component
+from data_extraction import Book
+from config import find_non_core_pages 
+from paths import CLEANED_TEXT, make_data_directories
+from segmentation import get_tokens_with_spacy, segment_with_spacy, add_spacy_pipeline_component
 
 
 def read_pdf(book: Book) -> Document:
